setting react props and children components to let the home page remain login
